requirements.txt
chardet==5.2.0
lxml==4.9.3
numpy==1.26.2
pandas==2.1.4
Pillow==10.1.0
PyPDF2==3.0.1
PyPDF4==1.27.0
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
reportlab==4.0.8
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.3
xlrd==2.0.1
